[db-slick] Create ConnectionIOEC phantom type for EC
This has several advantages: for one, will resolve conflicts with other
implicits when trying to do injection via the implicits mechanism.

Also, clearly state in the type what this EC is for. This is used when
doing map and flatmap over slick's DBIO, so there really shouldn't be
any weird side effects run in there, unless you really want to, I guess
:/ like christ.